refactor: update CharacterCreator component to use 8bit Button, enhance styling with retro class, and streamline genre selection logic
